增强集合框架试练
