--- ChapterIV/src/yingdg/exercise/collection/CollectionEx.java
@@ -1,16 +1,0 @@
-package yingdg.exercise.collection;
-
-import org.junit.Test;
-
-/**
- * Created by YingDG on 2017/2/4.
- */
-public class CollectionEx {
-    private String testStr = "123 abc 232 a2b4c3";
-
-    @Test
-    public void stringsTest() {
-
-    }
-
-}
